does this title work
i didn't know you could put a description, I'm so dumb. anyways, shooting works 100% now, but I couldn't get a consistent way for the projectiles to spawn with an offset (AND have the towers be accurate), so all towers shoot from the middle for now. maybe a slight offset is okay, but apparently bloons doesn't even have one